Added new methods to populate the lists of Subkeys and Entries of a RegKey object.
Added new methods to print contents of subkeys and entries of a registry
key. I then updated the old print function of RegKey accordingly and
removed the booleans. A list of the strings is now displayed from
printRegKey. printRegKey now prints full paths for parents and subkeys.

Commented out the netcatEntry method of RegEntry pending a likely decision
to change all print functions to return text which can then be piped
anywhere.

Updated the funciton passed to walkReg to accept a list as a parameter to
allow flexibility to the end-user.

Comment house-keeping. Fixed some typos, elaborated, etc.